requirements.txt
arrow==1.2.3
asgiref==3.6.0
binaryornot==0.4.4
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
Django~=4.1
django-browser-reload==1.6.0
django-extensions==3.2.1
django-taggit==3.1.0
django-tailwind==3.4.0
easy-thumbnails==2.8.5
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-slugify==7.0.0
python3-openid==3.2.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
urllib3==1.26.13
uwsgi==2.0.20
Werkzeug==2.2.2
wincertstore==0.2